# src/docs/swagger/collections.yaml
openapi: 3.0.0
info:
  title: Shopease
  description: The API collection for Shopease
  version: 1.0.0
servers:
  - url: http://localhost:9000/api/v1
    description: Shopease Development Server
  - url: http://api.shopease.shubhampurwar.in/api/v1
    description: Shopease Production Server
...
paths:
...
  /cart/add:
    post:
      tags:
        - Cart
      description: Allows users to add an item to their cart
      operationId: addItemToCart
      requestBody:
        description: ID of the item that is to be added to cart
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
      responses:
        '200':
          description: A JSON object containing a new cart item
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      product:
                        $ref: '#/components/schemas/Product'
                      quantity:
                        type: integer
  /cart/remove:
    put:
      tags:
        - Cart
      description: Allows users to remove an item from their cart
      operationId: removeItemFromCart
      requestBody:
        description: ID of the item that is to be removed from cart
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
      responses:
        '200':
          description: A JSON object containing ID of the removed item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ID'
  /cart/update:
    put:
      tags:
        - Cart
      description: Allows users to update quantity of a cart item
      operationId: updateItemQuantity
      requestBody:
        description: A JSON object containing product ID and quantity
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                quantity:
                  type: integer
      responses:
        '200':
          description: A JSON object containing a cart item whose quantity was updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      product:
                        $ref: '#/components/schemas/Product'
                      quantity:
                        type: integer
  /cart/move:
    put:
      tags:
        - Cart
      description: Alows users to move an item from their cart to wishlist
      operationId: moveItemToWishlist
      requestBody:
        description: ID of the item that is to be moved to wishlist
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
      responses:
        '200':
          description: A JSON object containing an item that was moved to wishlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Product'
...
  /categories/{categoryId}:
    get:
      tags:
        - Categories
      description: Fetches a category by ID
      operationId: getCategoryById
      parameters:
        - in: path
          name: categoryId
          required: true
          description: ID of the category to be fetched
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing a category
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Category'
  /admin/categories:
    post:
      tags:
        - Categories
      description: Allows admins to add a new category
      operationId: addNewCategory
      requestBody:
        description: Payload containing title and image of a category
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CategoryBody'
            encoding:
              image:
                contentType: image/*
      responses:
        '201':
          description: A JSON object containing a new category
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Category'
  /admin/categories/{categoryId}:
    post:
      tags:
        - Categories
      description: Allows admins to update a category
      operationId: updateCategory
      parameters:
        - in: path
          name: categoryId
          required: true
          description: ID of the category to be updated
          schema:
            type: string
      requestBody:
        description: Payload containing title and image of a category
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CategoryBody'
            encoding:
              image:
                contentType: image/*
      responses:
        '200':
          description: A JSON object containing an updated category
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Category'
...
components:
...
  schemas:
...
    ID:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            id:
              type: string
...
    Product:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        brand:
          type: string
        category:
          type: string
        stock:
          type: integer
        image:
          type: object
          properties:
            url:
              type: string
            publicId:
              type: string
        soldUnits:
          type: integer
        avgRating:
          type: number
        reviewCount:
          type: integer
        createdBy:
          type: string
        lastUpdatedBy:
          type: string
        deletedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        isDeleted:
          type: boolean
...
    Category:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        image:
          type: object
          properties:
            url:
              type: string
            publicId:
              type: string
        createdBy:
          type: string
        lastUpdatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategoryBody:
      type: object
      properties:
        image:
          type: string
          format: binary
        title:
          type: string
      required:
        - image
        - title
...